Web: Push messages down for banners and centralize filename truncation
- Move connection banners from an absolute overlay to the flow of the chat region so they push messages down instead of obscuring them.
- Adjust the top inset of the message list dynamically based on banner visibility using a shared CSS variable for the header height.
- Extract duplicated filename truncation logic into a central utility and link the header logo to the home route.
