+++ client/web/src/app/features/chat/components/chat-messages/chat-messages.component.html
@@ -1,5 +1,9 @@
 <!-- MARK: - Messages Container -->
-<div class="w-full flex-1 overflow-auto pt-[72px] min-h-0 no-scrollbar" #messageContainer>
+<div
+  class="w-full flex-1 overflow-auto min-h-0 no-scrollbar"
+  [style.paddingTop]="clearTopInset ? 'calc(var(--chat-bar-h) + 1rem)' : '0.5rem'"
+  #messageContainer
+>
   <div class="flex w-full flex-col items-start justify-start gap-4">
     <!-- MARK: - Welcome State (Empty Messages) -->
     @if (messages.length === 0) {
